added delete single sub-admin feature
